+++ src/main/java/com/alkemy/ong/data/gateways/DefaultSlidesGateway.java
@@ -2,12 +2,14 @@ package com.alkemy.ong.data.gateways;
 
 import com.alkemy.ong.data.entity.SlidesEntity;
 import com.alkemy.ong.data.repository.SlidesRepository;
-import com.alkemy.ong.domain.slides.gateway.SlidesGateway;
-import com.alkemy.ong.domain.slides.model.Slides;
+import com.alkemy.ong.domain.slides.SlidesGateway;
+import com.alkemy.ong.domain.slides.Slides;
 import org.springframework.stereotype.Component;
 
 import java.util.List;
-import java.util.stream.Collectors;
+
+import static java.time.LocalDateTime.now;
+import static java.util.stream.Collectors.toList;
 
 @Component
 public class DefaultSlidesGateway implements SlidesGateway {
