+++ src/main/java/com/alkemy/ong/domain/slides/Slides.java
@@ -1,4 +1,4 @@
-package com.alkemy.ong.domain.slides.model;
+package com.alkemy.ong.domain.slides;
 
 import lombok.*;
